% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ArrayType}
\alias{ArrayType}
\title{FFI Array Type}
\usage{
ArrayType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  element_type = FFIType(),
  length = integer(0)
)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}

\item{element_type}{FFIType of array elements}

\item{length}{Integer length of array}
}
\value{
An ArrayType object
}
\description{
FFI Array Type
}
\keyword{Types}
