% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindgen_r_api.R
\name{bindgen_r_api_summary}
\alias{bindgen_r_api_summary}
\title{Get summary of R API bindings}
\usage{
bindgen_r_api_summary(include_path = R.home("include"))
}
\arguments{
\item{include_path}{Path to R's include directory}
}
\value{
Data frame with header info
}
\description{
Quick summary of what's available in R's C API headers.
}
\examples{
\dontrun{
bindgen_r_api_summary()
}
}
