% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{CIF}
\alias{CIF}
\title{FFI Call Interface (CIF)}
\usage{
CIF(return_type = FFIType(), arg_types = list(), ref = NULL)
}
\arguments{
\item{return_type}{FFIType for return value}

\item{arg_types}{List of FFIType objects for arguments}

\item{ref}{External pointer to ffi_cif}
}
\value{
An CIF object
}
\description{
FFI Call Interface (CIF)
}
\keyword{Types}
