% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{create_builtin_type}
\alias{create_builtin_type}
\title{Create built-in FFI type}
\usage{
create_builtin_type(name, ...)
}
\arguments{
\item{name}{Character name of built-in type}
}
\description{
Create built-in FFI type
}
