% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dll_loading.R
\name{dll_info}
\alias{dll_info}
\title{Get information about a loaded library}
\usage{
dll_info(handle)
}
\arguments{
\item{handle}{Library handle (path)}
}
\value{
List with library information
}
\description{
Get information about a loaded library
}
