% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dll_loading.R, R/dll_loading_broken.R
\name{dll_is_loaded}
\alias{dll_is_loaded}
\title{Check if a symbol is loaded}
\usage{
dll_is_loaded(symbol_name, package = NULL)

dll_is_loaded(symbol_name, package = NULL)
}
\arguments{
\item{symbol_name}{Name of the symbol to check}

\item{package}{Package name (optional)}
}
\value{
TRUE if symbol is loaded, FALSE otherwise

TRUE if symbol is loaded, FALSE otherwise
}
\description{
Check if a symbol is loaded

Check if a symbol is loaded
}
