% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dll_loading.R, R/dll_loading_broken.R
\name{dll_symbol}
\alias{dll_symbol}
\title{Get symbol information from a loaded library}
\usage{
dll_symbol(symbol_name, package = NULL)

dll_symbol(symbol_name, package = NULL)
}
\arguments{
\item{symbol_name}{Name of the symbol to find}

\item{package}{Package name where symbol is registered (optional)}
}
\value{
Symbol information including address as external pointer

Symbol information including address as external pointer
}
\description{
Get symbol information from a loaded library

Get symbol information from a loaded library
}
