% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dll_loading.R
\name{dll_unload}
\alias{dll_unload}
\title{Unload a shared library/DLL}
\usage{
dll_unload(handle, verbose = FALSE)
}
\arguments{
\item{handle}{Library handle (path) returned by dll_load()}

\item{verbose}{Print unloading information (default FALSE)}
}
\description{
Unload a shared library/DLL
}
