% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindgen_r_api.R
\name{.find_r_library}
\alias{.find_r_library}
\title{Find R shared library path}
\usage{
.find_r_library()
}
\value{
Path to R shared library or NULL if not found
}
\description{
Find R shared library path
}
\keyword{internal}
