% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{EnumType}
\alias{EnumType}
\title{FFI Enumeration Type}
\usage{
EnumType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  values = integer(0),
  underlying_type = FFIType()
)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}

\item{values}{Named integer vector of enum values}

\item{underlying_type}{FFIType for the underlying integer type}
}
\description{
FFI Enumeration Type
}
