% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-structures.R
\name{ffi_alloc}
\alias{ffi_alloc}
\title{Allomessagee memory for FFI structure}
\usage{
ffi_alloc(type, ...)
}
\arguments{
\item{type}{StructType object}
}
\description{
Allomessagee memory for FFI structure
}
