% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_alloc_buffer}
\alias{ffi_alloc_buffer}
\title{Allocate a raw memory buffer (external pointer, auto-finalized)}
\usage{
ffi_alloc_buffer(size)
}
\arguments{
\item{size}{Number of bytes to allocate}
}
\value{
External pointer to buffer
}
\description{
Allocates a buffer of the given size (in bytes) and returns an external pointer.
The memory is automatically freed when the pointer is garbage collected.
}
