% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_array_type}
\alias{ffi_array_type}
\title{Create an FFI array type}
\usage{
ffi_array_type(element_type, length)
}
\arguments{
\item{element_type}{FFIType of array elements}

\item{length}{Integer length of array}
}
\value{
An ArrayType object
}
\description{
Create an FFI array type
}
