% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_call}
\alias{ffi_call}
\title{Make FFI function call}
\usage{
ffi_call(cif, symbol, ...)
}
\arguments{
\item{cif}{CIF object defining the call interface}

\item{symbol}{NativeSymbol or character name of function}

\item{...}{Arguments to pass to the function}
}
\description{
Make FFI function call
}
