% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_cif}
\alias{ffi_cif}
\title{Prepare FFI call interface
Prepare FFI call interface}
\usage{
ffi_cif(return_type, ...)
}
\arguments{
\item{return_type}{FFIType for return value}

\item{...}{FFIType objects for arguments}
}
\description{
Prepare FFI call interface
Prepare FFI call interface
}
