% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_helpers.R
\name{ffi_cleanup_lib}
\alias{ffi_cleanup_lib}
\title{Clean up compiled library}
\usage{
ffi_cleanup_lib(lib)
}
\arguments{
\item{lib}{Compiled library from ffi_compile_shlib()}
}
\value{
NULL (invisible)
}
\description{
Unload library and remove temporary files.
}
\keyword{internal}
