% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_closure_pointer}
\alias{ffi_closure_pointer}
\title{Get the function pointer for an FFI closure}
\usage{
ffi_closure_pointer(closure)
}
\arguments{
\item{closure}{An FFIClosure object}
}
\value{
External pointer to the callable function
}
\description{
Returns the executable function pointer that can be passed to C functions
expecting a callback.
}
\details{
The returned pointer can be passed to C functions via \code{ffi_call()}.
It will invoke the R function when called.
}
\seealso{
\link[RSimpleFFI:ffi_closure]{ffi_closure()} to create closures
}
