% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_closures_supported}
\alias{ffi_closures_supported}
\title{Check if closures are supported on this platform}
\usage{
ffi_closures_supported()
}
\value{
Logical; TRUE if closures are supported
}
\description{
Not all platforms support FFI closures. Use this function to check
before attempting to create closures.
}
