% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_copy_array}
\alias{ffi_copy_array}
\title{Copy array from native memory}
\usage{
ffi_copy_array(ptr, length, element_type)
}
\arguments{
\item{ptr}{External pointer to array}

\item{length}{Integer length of array}

\item{element_type}{FFIType of array elements}
}
\description{
Copy array from native memory
}
