% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_copy_array_type}
\alias{ffi_copy_array_type}
\title{Copy array from native memory (ArrayType version)}
\usage{
ffi_copy_array_type(ptr, array_type)
}
\arguments{
\item{ptr}{External pointer to array}

\item{array_type}{ArrayType object}
}
\description{
Copy array from native memory (ArrayType version)
}
