% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_deref_pointer}
\alias{ffi_deref_pointer}
\title{Dereference a pointer}
\usage{
ffi_deref_pointer(ptr)
}
\arguments{
\item{ptr}{External pointer to the address to dereference}
}
\value{
External pointer containing the value at the address
}
\description{
Reads the pointer value stored at an address. This is useful for
accessing global variables in shared libraries that are pointers
(like R_GlobalEnv, R_NilValue, etc.).
}
\examples{
\dontrun{
# Get R_GlobalEnv from libR.so
addr <- getNativeSymbolInfo("R_GlobalEnv")$address
globalenv_sexp <- ffi_deref_pointer(addr)
}
}
