% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_enum}
\alias{ffi_enum}
\title{Create FFI enumeration type}
\usage{
ffi_enum(..., underlying_type = ffi_int())
}
\arguments{
\item{...}{Named integer values representing enum constants}

\item{underlying_type}{FFIType for underlying integer type (default: ffi_int())}
}
\value{
EnumType object
}
\description{
Create FFI enumeration type
}
\keyword{Types}
