% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffi_enum_to_int}
\alias{ffi_enum_to_int}
\title{Convert enum name to integer value}
\usage{
ffi_enum_to_int(enum_type, name)
}
\arguments{
\item{enum_type}{EnumType object}

\item{name}{Character name of enum constant}
}
\value{
Integer value
}
\description{
Look up the integer value for a named enum constant.
}
\examples{
\dontrun{
Color <- ffi_enum(RED = 0L, GREEN = 1L, BLUE = 2L)
ffi_enum_to_int(Color, "GREEN") # 1L
}
}
