% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfield_helpers.R
\name{ffi_extract_bits64}
\alias{ffi_extract_bits64}
\title{Extract a single bit-field from a 64-bit packed value}
\usage{
ffi_extract_bits64(packed_value, bit_offset, bit_width)
}
\arguments{
\item{packed_value}{Packed value (as double for 64-bit range)}

\item{bit_offset}{Bit offset from LSB (0-based)}

\item{bit_width}{Number of bits in the field}
}
\value{
Extracted value as double (for 64-bit range)
}
\description{
Extract a single bit-field from a 64-bit packed value
}
\examples{
packed <- ffi_pack_bits64(c(1L, 5L, 12L), c(1L, 3L, 4L))
ffi_extract_bits64(packed, 1L, 3L) # 5 (mode field)

}
