% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mode.R
\name{ffi_field_to_r}
\alias{ffi_field_to_r}
\title{Convert field pointer to R value}
\usage{
ffi_field_to_r(field_ptr)
}
\arguments{
\item{field_ptr}{External pointer to field (must be tagged with FFI type)}
}
\value{
Field value converted to appropriate R type
}
\description{
Converts a tagged field pointer to an R value using the type information
stored in the pointer's tag.
}
\keyword{internal}
