% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_fill_typed_buffer}
\alias{ffi_fill_typed_buffer}
\title{Fill a typed buffer from an R vector (int or double)}
\usage{
ffi_fill_typed_buffer(ptr, values, type)
}
\arguments{
\item{ptr}{External pointer to buffer}

\item{values}{Integer or double vector}

\item{type}{FFIType object}
}
\description{
Fill a typed buffer from an R vector (int or double)
}
