% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R, R/api_mode.R
\name{ffi_get_field}
\alias{ffi_get_field}
\title{Get field value from FFI structure}
\usage{
ffi_get_field(struct_ptr, field_name, struct_type)

ffi_get_field(struct_ptr, field_name, struct_type)
}
\arguments{
\item{struct_ptr}{External pointer to the struct}

\item{field_name}{Name of the field to get}

\item{struct_type}{Struct type metadata (contains field offsets and types)}

\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}

\item{...}{Additional arguments (not used)}
}
\value{
Field value

Field value converted to appropriate R type
}
\description{
Reads a field value using type-aware conversion. Works with bitfields, regular
fields, pointers, nested structs, etc.
}
\keyword{Types}
\keyword{internal}
