% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_helpers.R
\name{ffi_get_symbol}
\alias{ffi_get_symbol}
\title{Get function pointer from compiled library}
\usage{
ffi_get_symbol(lib, symbol_name)
}
\arguments{
\item{lib}{Compiled library from ffi_compile_shlib()}

\item{symbol_name}{Name of the C function}
}
\value{
Function that calls the C function via .Call()
}
\description{
Extract a function pointer from a compiled library for use with .Call().
}
\keyword{internal}
