% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffi_is_null}
\alias{ffi_is_null}
\title{Check if external pointer is NULL}
\usage{
ffi_is_null(ptr)
}
\arguments{
\item{ptr}{External pointer}
}
\value{
Logical
}
\description{
Check if external pointer is NULL
}
