% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffi_print_struct}
\alias{ffi_print_struct}
\title{Pretty print struct contents}
\usage{
ffi_print_struct(ptr, struct_type)
}
\arguments{
\item{ptr}{External pointer to struct}

\item{struct_type}{StructType object}
}
\description{
Pretty print struct contents
}
