% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{ffi_read_global}
\alias{ffi_read_global}
\title{Read a global variable from a shared library}
\usage{
ffi_read_global(ptr, type)
}
\arguments{
\item{ptr}{External pointer to the global variable address}

\item{type}{FFIType describing the type of the global variable}
}
\value{
The value at the address, converted to an appropriate R type
}
\description{
Reads a typed value from a global symbol address. This is a typed
version of ffi_deref_pointer that handles type conversion.
}
