% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfield_helpers.R
\name{ffi_set_bits64}
\alias{ffi_set_bits64}
\title{Set a single bit-field in a 64-bit packed value}
\usage{
ffi_set_bits64(packed_value, new_value, bit_offset, bit_width)
}
\arguments{
\item{packed_value}{Packed value (as double for 64-bit range)}

\item{new_value}{New value for the bit-field}

\item{bit_offset}{Bit offset from LSB (0-based)}

\item{bit_width}{Number of bits in the field}
}
\value{
Modified packed value as double
}
\description{
Set a single bit-field in a 64-bit packed value
}
\examples{
packed <- ffi_pack_bits64(c(1L, 5L, 12L), c(1L, 3L, 4L))
new_packed <- ffi_set_bits64(packed, 7L, 1L, 3L)  # Set mode to 7
ffi_extract_bits64(new_packed, 1L, 3L)  # 7

}
