% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_sizeof}
\alias{ffi_sizeof}
\title{Get size of FFI type in bytes}
\usage{
ffi_sizeof(type, ...)
}
\arguments{
\item{type}{FFIType object}

\item{...}{Additional arguments (not used)}
}
\value{
Size in bytes
}
\description{
Get size of FFI type in bytes
}
\section{Types}{
NA
}

