% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffi_struct_array_from_list}
\alias{ffi_struct_array_from_list}
\title{Allocate array of structs from R list}
\usage{
ffi_struct_array_from_list(struct_type, values)
}
\arguments{
\item{struct_type}{StructType object}

\item{values}{List of named lists, one per struct}
}
\value{
External pointer to allocated struct array
}
\description{
Allocate array of structs from R list
}
\examples{
\dontrun{
Point <- ffi_struct(x = ffi_int(), y = ffi_int())
points <- ffi_struct_array_from_list(Point, list(
  list(x = 0L, y = 0L),
  list(x = 10L, y = 20L),
  list(x = 30L, y = 40L)
))
}
}
