% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffi_struct_to_list}
\alias{ffi_struct_to_list}
\title{Convert struct to R list}
\usage{
ffi_struct_to_list(ptr, struct_type)
}
\arguments{
\item{ptr}{External pointer to struct}

\item{struct_type}{StructType object}
}
\value{
Named list of field values
}
\description{
Convert struct to R list
}
\examples{
\dontrun{
Point <- ffi_struct(x = ffi_int(), y = ffi_int())
pt <- ffi_alloc(Point)
ffi_set_field(pt, "x", 42L, Point)
ffi_set_field(pt, "y", 100L, Point)
as.list(pt, Point) # list(x = 42L, y = 100L)
}
}
