% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_symbol}
\alias{ffi_symbol}
\title{Get native symbol reference}
\usage{
ffi_symbol(name, library = NULL)
}
\arguments{
\item{name}{Character name of the symbol}

\item{library}{Character name of library (optional)}
}
\description{
Get native symbol reference
}
