% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_symbol_from_address}
\alias{ffi_symbol_from_address}
\title{Create native symbol from direct address}
\usage{
ffi_symbol_from_address(address, name = "anonymous")
}
\arguments{
\item{address}{External pointer to the symbol address}

\item{name}{Character name of the symbol (for reference)}
}
\description{
Create native symbol from direct address
}
