% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfield_helpers.R
\name{ffi_unpack_bits64}
\alias{ffi_unpack_bits64}
\title{Unpack bit-fields from a 64-bit value}
\usage{
ffi_unpack_bits64(packed_value, widths)
}
\arguments{
\item{packed_value}{Packed value (as double for 64-bit range)}

\item{widths}{Integer vector of bit widths for each field}
}
\value{
Integer vector of unpacked values
}
\description{
Extracts multiple values from a packed 64-bit value according to specified
bit widths.
}
\examples{
packed <- ffi_pack_bits64(c(1L, 5L, 12L), c(1L, 3L, 4L))
ffi_unpack_bits64(packed, c(1L, 3L, 4L)) # c(1, 5, 12)

}
