% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{FFIClosure}
\alias{FFIClosure}
\title{FFI Closure - R function as C callback}
\usage{
FFIClosure(
  r_function = function() NULL,
  cif = CIF(),
  ref = NULL,
  func_ptr = NULL
)
}
\arguments{
\item{r_function}{The R function to wrap}

\item{cif}{CIF object describing the callback signature}

\item{ref}{External pointer to the closure}

\item{func_ptr}{External pointer to the executable function}
}
\value{
An FFIClosure object
}
\description{
A closure wraps an R function so it can be used as a callback
from C code. The closure has an associated CIF that describes
the function signature.
}
\keyword{Types}
