% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\docType{package}
\name{FFIType}
\alias{-package}
\alias{FFIType}
\title{FFI Type representation
FFI Type representation}
\usage{
FFIType(name = character(0), size = integer(0), ref = NULL)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}
}
\keyword{internal}
