% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{FieldInfo}
\alias{FieldInfo}
\title{Field Information Class}
\usage{
FieldInfo(
  name = character(0),
  type = FFIType(),
  offset = integer(0),
  size = integer(0),
  index = integer(0)
)
}
\description{
Represents metadata about a single field in a structure.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character name of the field}

\item{\code{type}}{FFIType of the field}

\item{\code{offset}}{Integer byte offset within structure}

\item{\code{size}}{Integer size of field in bytes}

\item{\code{index}}{Integer 1-based field index}
}}

