% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_api_mode.R
\name{generate_api_r_wrappers}
\alias{generate_api_r_wrappers}
\title{Generate R wrapper code that calls compiled C accessors}
\usage{
generate_api_r_wrappers(
  struct_name,
  field_names,
  prefix = "rffi_",
  package_name = NULL
)
}
\arguments{
\item{struct_name}{Name of the struct}

\item{field_names}{Character vector of field names}

\item{prefix}{Prefix for C function names (default "rffi_")}

\item{package_name}{Name of the package (for .Call interface)}
}
\value{
Character string containing R code
}
\description{
Creates R functions that use .Call() to invoke compiled struct helpers.
}
\keyword{internal}
