% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generation.R
\name{generate_api_struct_typedef}
\alias{generate_api_struct_typedef}
\title{Generate struct typedef from field types}
\usage{
generate_api_struct_typedef(struct_name, field_names, field_types = NULL)
}
\arguments{
\item{struct_name}{Name of the C struct}

\item{field_names}{Character vector of field names}

\item{field_types}{Character vector of C type names (e.g., "int", "double")}
}
\value{
Character string containing typedef
}
\description{
Generate struct typedef from field types
}
\keyword{internal}
