% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_bitfield_accessor_code}
\alias{generate_bitfield_accessor_code}
\title{Generate bit-field accessor code}
\usage{
generate_bitfield_accessor_code(struct_name, bitfield_specs)
}
\arguments{
\item{struct_name}{Name of the struct}

\item{bitfield_specs}{Character vector of "'name : width'" strings}
}
\value{
Character string with accessor code
}
\description{
Generate bit-field accessor code
}
\keyword{internal}
