% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_enum_definition}
\alias{generate_enum_definition}
\title{Generate R enum definition from parsed enum}
\usage{
generate_enum_definition(enum_name, enum_values)
}
\arguments{
\item{enum_name}{Name of the enum}

\item{enum_values}{Named integer vector of enum values}
}
\value{
Character vector with R code
}
\description{
Generate R enum definition from parsed enum
}
