% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_function_wrapper}
\alias{generate_function_wrapper}
\title{Generate R function wrapper from parsed function}
\usage{
generate_function_wrapper(func_def)
}
\arguments{
\item{func_def}{Function definition (row from functions data.frame)}
}
\value{
Character vector with R code
}
\description{
Generate R function wrapper from parsed function
}
