% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_struct_definition}
\alias{generate_struct_definition}
\title{Generate R struct definition from parsed struct}
\usage{
generate_struct_definition(struct_name, struct_def)
}
\arguments{
\item{struct_name}{Name of the struct}

\item{struct_def}{Struct definition from parsed header}
}
\value{
Character vector with R code
}
\description{
Generate R struct definition from parsed struct
}
