% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_union_definition}
\alias{generate_union_definition}
\title{Generate R union definition from parsed union}
\usage{
generate_union_definition(union_name, union_def)
}
\arguments{
\item{union_name}{Name of the union}

\item{union_def}{Union definition (list of fields)}
}
\value{
Character vector with R code
}
\description{
Generate R union definition from parsed union
}
