% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{generate_wrapper_code}
\alias{generate_wrapper_code}
\title{Generate wrapper code from parameter information}
\usage{
generate_wrapper_code(
  func_name,
  return_type_ffi,
  param_names,
  param_types_c,
  param_types_ffi,
  is_variadic
)
}
\description{
Generate wrapper code from parameter information
}
\keyword{internal}
