% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-utils.R
\name{get_pointer_type}
\alias{get_pointer_type}
\title{Get pointer type tag}
\usage{
get_pointer_type(ptr)
}
\arguments{
\item{ptr}{External pointer}
}
\value{
Character string with the type name
}
\description{
Retrieves the type tag from an external pointer.
}
