% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing_codegen.R
\name{get_resolvable_types}
\alias{get_resolvable_types}
\title{Get list of resolvable type names}
\usage{
get_resolvable_types()
}
\value{
Character vector of resolvable type names
}
\description{
Returns a character vector of all type names that can be directly resolved
to FFI types, for use in can_resolve_typedef().
}
\keyword{internal}
