% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexp_helpers.R
\name{is_protected_ptr}
\alias{is_protected_ptr}
\title{Check if an object is a protected SEXP pointer}
\usage{
is_protected_ptr(ptr)
}
\arguments{
\item{ptr}{An external pointer}
}
\value{
TRUE if ptr was created by sexp_ptr() or data_ptr()
}
\description{
Check if an object is a protected SEXP pointer
}
