% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_parsing.R
\name{map_type_to_ffi}
\alias{map_type_to_ffi}
\title{Map C type to FFI type string}
\usage{
map_type_to_ffi(type_string, type_map, typedef_ffi_map, is_return = FALSE)
}
\description{
Map C type to FFI type string
}
\keyword{internal}
