% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{NativeSymbol}
\alias{NativeSymbol}
\title{Native Symbol Reference}
\usage{
NativeSymbol(name = character(0), address = NULL, library = character(0))
}
\arguments{
\item{name}{Character name of the symbol}

\item{address}{External pointer to the symbol}

\item{library}{Character name of library (optional)}
}
\description{
Native Symbol Reference
}
