% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_api_mode.R
\name{package_api_mode}
\alias{package_api_mode}
\title{API Mode Package Generation Functions}
\description{
Functions for generating R packages with compiled struct helpers (API mode)
}
\keyword{internal}
