% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexp_helpers.R
\name{ptr_to_sexp}
\alias{ptr_to_sexp}
\title{Get the R object from a protected SEXP pointer}
\usage{
ptr_to_sexp(ptr)
}
\arguments{
\item{ptr}{A protected SEXP pointer}
}
\value{
The original R object
}
\description{
Retrieves the original R object from a pointer created by sexp_ptr().
}
