% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{regexec_escape}
\alias{regexec_escape}
\title{Escape a string for use in a regular expression}
\usage{
regexec_escape(x)
}
\description{
Escape a string for use in a regular expression
}
\keyword{internal}
