% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{RSimpleFFI-package}
\alias{RSimpleFFI}
\alias{RSimpleFFI-package}
\title{RSimpleFFI: Simple Foreign Function Interface using 'S7' and 'libffi'}
\description{
Simple Foreign Function Interface for 'R' using 'libffi' and 'S7' classes. Supports calling 'C' functions with type conversion and struct handling. Includes standard 'C' types (int8, int16, int32, int64, uint variants), platform types (size_t, bool), floating point types, and complex struct types. Experimental header parsing using 'tinycc' enables automatic generation of 'R' bindings from 'C' header files, simplifying package development for 'C' libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/RSimpleFFI}
  \item \url{https://sounkou-bioinfo.github.io/RSimpleFFI/}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/RSimpleFFI/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Anthony Green, Red Hat, Inc and others (libffi authors and COPYRIGHT holders) [contributor]
  \item Fabrice Bellard and tinycc Authors (Tinycc Compiler (tinycc) authors and COPYRIGHT holders) [contributor]
}

}
\keyword{internal}
