% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{StructType}
\alias{StructType}
\title{FFI Structure Type}
\usage{
StructType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  fields = character(0),
  field_types = list()
)
}
\arguments{
\item{fields}{Character vector of field names}

\item{field_types}{List of FFIType objects for each field}
}
\description{
FFI Structure Type
}
