% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_binary_path}
\alias{tcc_binary_path}
\title{Get path to embedded TCC binary}
\usage{
tcc_binary_path()
}
\value{
Path to tcc executable in installed package
}
\description{
Get path to embedded TCC binary
}
