% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_extract_defines}
\alias{tcc_extract_defines}
\title{Extract #define macros from C header file}
\usage{
tcc_extract_defines(header_file)
}
\arguments{
\item{header_file}{Path to C header file}
}
\value{
Named list of macro definitions
}
\description{
Extract #define macros from C header file
}
