% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_extract_structs}
\alias{tcc_extract_structs}
\title{Extract struct definitions from preprocessed C code}
\usage{
tcc_extract_structs(preprocessed_lines)
}
\arguments{
\item{preprocessed_lines}{Character vector from tcc_preprocess()}
}
\value{
List of struct definitions (name -> list of fields).
Each struct may have a "packed" attribute set to TRUE if \strong{attribute}((packed))
or #pragma pack was detected.
}
\description{
Extract struct definitions from preprocessed C code
}
