% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_extract_unions}
\alias{tcc_extract_unions}
\title{Extract union definitions from preprocessed C code}
\usage{
tcc_extract_unions(preprocessed_lines)
}
\arguments{
\item{preprocessed_lines}{Character vector from tcc_preprocess()}
}
\value{
List of union definitions (name -> list of fields)
}
\description{
Extract union definitions from preprocessed C code
}
