% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_parse_header}
\alias{tcc_parse_header}
\title{Parse C header file and extract all declarations}
\usage{
tcc_parse_header(header_file, includes = NULL)
}
\arguments{
\item{header_file}{Path to C header file}

\item{includes}{Additional include directories}
}
\value{
List with components: functions, structs, defines
}
\description{
Parse C header file and extract all declarations
}
