% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_preprocess}
\alias{tcc_preprocess}
\title{Preprocess C header file using embedded TCC}
\usage{
tcc_preprocess(header_file, includes = NULL, keep_defines = FALSE)
}
\arguments{
\item{header_file}{Path to C header file}

\item{includes}{Additional include directories}

\item{keep_defines}{Keep #define directives (not supported by -E alone)}
}
\value{
Character vector of preprocessed lines
}
\description{
Preprocess C header file using embedded TCC
}
