% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_wrapper.R
\name{tcc_run}
\alias{tcc_run}
\title{Compile and run C code using embedded TCC}
\usage{
tcc_run(code, args = character())
}
\arguments{
\item{code}{C source code as string}

\item{args}{Arguments to pass to compiled program}
}
\value{
Output from program
}
\description{
Compile and run C code using embedded TCC
}
