% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_check_packed_attribute}
\alias{ts_check_packed_attribute}
\title{Check if a struct has packed attribute}
\usage{
ts_check_packed_attribute(struct_node, source_text)
}
\description{
Check if a struct has packed attribute
}
\keyword{internal}
