% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_count_pointer_stars}
\alias{ts_count_pointer_stars}
\title{Count asterisks in a pointer_declarator by traversing AST}
\usage{
ts_count_pointer_stars(declarator_node)
}
\arguments{
\item{declarator_node}{A pointer_declarator node}
}
\value{
Integer count of pointer asterisks
}
\description{
Counts the number of pointer indirection levels by finding all '*' nodes
in the pointer_declarator subtree.
}
\keyword{internal}
