% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_extract_enums}
\alias{ts_extract_enums}
\title{Extract enum definitions from tree-sitter AST}
\usage{
ts_extract_enums(root_node, source_text)
}
\description{
Extract enum definitions from tree-sitter AST
}
\keyword{internal}
