% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_extract_functions}
\alias{ts_extract_functions}
\title{Extract function declarations from tree-sitter AST}
\usage{
ts_extract_functions(root_node, source_text)
}
\description{
Extract function declarations from tree-sitter AST
}
\keyword{internal}
