% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_extract_struct_fields}
\alias{ts_extract_struct_fields}
\title{Extract fields from a struct body node}
\usage{
ts_extract_struct_fields(body_node, source_text)
}
\description{
Extract fields from a struct body node
}
\keyword{internal}
