% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_extract_typedefs}
\alias{ts_extract_typedefs}
\title{Extract typedef declarations from tree-sitter AST}
\usage{
ts_extract_typedefs(root_node, source_text)
}
\description{
Extract typedef declarations from tree-sitter AST
}
\keyword{internal}
