% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_extract_unions}
\alias{ts_extract_unions}
\title{Extract union definitions from tree-sitter AST}
\usage{
ts_extract_unions(root_node, source_text)
}
\description{
Extract union definitions from tree-sitter AST
}
\keyword{internal}
