% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_find_function_declarator}
\alias{ts_find_function_declarator}
\title{Find function_declarator recursively inside pointer_declarator chain}
\usage{
ts_find_function_declarator(node)
}
\arguments{
\item{node}{A pointer_declarator node}
}
\value{
The function_declarator node if found, NULL otherwise
}
\description{
For declarations like \code{int** foo(...)}, the function_declarator is nested
inside multiple pointer_declarator nodes.
}
\keyword{internal}
