% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_get_array_dimensions}
\alias{ts_get_array_dimensions}
\title{Get array dimensions from an array declarator}
\usage{
ts_get_array_dimensions(declarator_node, source_text)
}
\description{
Get array dimensions from an array declarator
}
\keyword{internal}
