% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_get_declarator_name}
\alias{ts_get_declarator_name}
\title{Get the name from a declarator node}
\usage{
ts_get_declarator_name(declarator_node, source_text)
}
\description{
Get the name from a declarator node
}
\keyword{internal}
