% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_get_typedef_info}
\alias{ts_get_typedef_info}
\title{Get typedef name and pointer count from declarator}
\usage{
ts_get_typedef_info(declarator_node, source_text)
}
\description{
Get typedef name and pointer count from declarator
}
\keyword{internal}
