% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_get_typedef_name}
\alias{ts_get_typedef_name}
\title{Get typedef name from declarator (legacy wrapper)}
\usage{
ts_get_typedef_name(declarator_node, source_text)
}
\description{
Get typedef name from declarator (legacy wrapper)
}
\keyword{internal}
