% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_array_declarator}
\alias{ts_parse_array_declarator}
\title{Parse array declarator to extract array suffix and identifier}
\usage{
ts_parse_array_declarator(array_node, source_text)
}
\description{
Parse array declarator to extract array suffix and identifier
}
\keyword{internal}
