% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_enum_values}
\alias{ts_parse_enum_values}
\title{Parse enum values from enumerator_list node}
\usage{
ts_parse_enum_values(body_node, source_text)
}
\description{
Parse enum values from enumerator_list node
}
\keyword{internal}
