% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_enumerator}
\alias{ts_parse_enumerator}
\title{Parse a single enumerator node}
\usage{
ts_parse_enumerator(enum_node, source_text, default_value)
}
\description{
Parse a single enumerator node
}
\keyword{internal}
