% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_field_declaration}
\alias{ts_parse_field_declaration}
\title{Parse a field declaration node}
\usage{
ts_parse_field_declaration(field_node, source_text)
}
\description{
Parse a field declaration node
}
\keyword{internal}
