% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_function_declarator}
\alias{ts_parse_function_declarator}
\title{Parse function declarator node}
\usage{
ts_parse_function_declarator(declarator_node, source_text, return_type)
}
\description{
Parse function declarator node
}
\keyword{internal}
