% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_parser.R
\name{ts_parse_parameter_declaration}
\alias{ts_parse_parameter_declaration}
\title{Parse a single parameter declaration}
\usage{
ts_parse_parameter_declaration(param_node, source_text)
}
\description{
Parse a single parameter declaration
}
\keyword{internal}
