% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{FFIType}
\alias{FFIType}
\alias{ArrayType}
\alias{ffi_array_type}
\alias{CIF}
\alias{NativeSymbol}
\alias{create_builtin_type}
\alias{ffi_raw}
\alias{ffi_void}
\alias{ffi_int}
\alias{ffi_double}
\alias{ffi_float}
\alias{ffi_pointer}
\alias{ffi_string}
\alias{ffi_int8}
\alias{ffi_int16}
\alias{ffi_int32}
\alias{ffi_int64}
\alias{ffi_char}
\alias{ffi_short}
\alias{ffi_long}
\alias{ffi_longlong}
\alias{ffi_uint}
\alias{ffi_uint8}
\alias{ffi_uint16}
\alias{ffi_uint32}
\alias{ffi_uint64}
\alias{ffi_uchar}
\alias{ffi_ushort}
\alias{ffi_ulong}
\alias{ffi_ulonglong}
\alias{ffi_longdouble}
\alias{ffi_size_t}
\alias{ffi_ssize_t}
\alias{ffi_bool}
\alias{ffi_wchar_t}
\alias{ffi_struct}
\alias{ffi_get_field}
\alias{ffi_set_field}
\alias{ffi_sizeof}
\title{FFI Type representation}
\usage{
FFIType(name = character(0), size = integer(0), ref = NULL)

ArrayType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  element_type = FFIType(),
  length = integer(0)
)

ffi_array_type(element_type, length)

CIF(return_type = FFIType(), arg_types = list(), ref = NULL)

NativeSymbol(name = character(0), address = NULL, library = character(0))

create_builtin_type(name, ...)

ffi_raw()

ffi_void()

ffi_int()

ffi_double()

ffi_float()

ffi_pointer()

ffi_string()

ffi_int8()

ffi_int16()

ffi_int32()

ffi_int64()

ffi_char()

ffi_short()

ffi_long()

ffi_longlong()

ffi_uint()

ffi_uint8()

ffi_uint16()

ffi_uint32()

ffi_uint64()

ffi_uchar()

ffi_ushort()

ffi_ulong()

ffi_ulonglong()

ffi_longdouble()

ffi_size_t()

ffi_ssize_t()

ffi_bool()

ffi_wchar_t()

ffi_struct(...)

ffi_get_field(ptr, field, struct_type, ...)

ffi_set_field(ptr, field, value, struct_type, ...)

ffi_sizeof(type, ...)
}
\arguments{
\item{name}{Character name of built-in type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_cif}

\item{element_type}{FFIType of array elements}

\item{length}{Integer length of array}

\item{return_type}{FFIType for return value}

\item{arg_types}{List of FFIType objects for arguments}

\item{address}{External pointer to the symbol}

\item{library}{Character name of library (optional)}

\item{...}{Additional arguments (not used)}

\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}

\item{struct_type}{StructType object}

\item{value}{Value to set}

\item{type}{FFIType object}
}
\value{
An FFIType object

An ArrayType object

An ArrayType object

An CIF object

A NativeSymbol object

An FFIType object

FFIType object for bool

FFIType object for char

FFIType object for double

FFIType object for float

FFIType object for int

FFIType object for int16

FFIType object for pointer

FFIType object for string

FFIType object for int8

FFIType object for int16

FFIType object for int32

FFIType object for int64

FFIType char type

FFIType object for short

FFIType object for long

FFIType object for longlong

FFIType object for ssize_t

FFIType object for uint8

FFIType object for uint16

FFIType object for uint32

FFIType object for uint64

FFIType object for uchar

FFIType object for ushort

FFIType object for ulong

FFIType object for ulonglong

FFIType object for longdouble

FFIType object for size_t

FFIType object for ssize_t

FFIType object for bool

FFIType object for wchar_t

StructType object

Field value

Updated pointer

Size in bytes
}
\description{
Wide char FFI type
}
