% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{UnionType}
\alias{UnionType}
\title{FFI Union Type}
\usage{
UnionType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  fields = character(0),
  field_types = list()
)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}

\item{fields}{Character vector of field names}

\item{field_types}{List of FFIType objects for each field}
}
\description{
FFI Union Type
}
