% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ArrayType}
\alias{ArrayType}
\title{FFI Array Type}
\usage{
ArrayType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  element_type = FFIType(),
  length = integer(0)
)
}
\arguments{
\item{element_type}{FFIType of array elements}

\item{length}{Integer length of array}
}
\description{
FFI Array Type
}
