% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexp_helpers.R
\name{data_ptr}
\alias{data_ptr}
\title{Get data pointer from R vector with GC protection}
\usage{
data_ptr(x)
}
\arguments{
\item{x}{An R vector (integer, double, complex, character, raw, or list)}
}
\value{
External pointer to the data, with finalizer to release protection
}
\description{
Returns an external pointer to the underlying data (INTEGER, REAL, etc.)
while ensuring the object won't be garbage collected.
}
\examples{
\dontrun{
x <- as.double(1:10)
ptr <- data_ptr(x)
# ptr points to the double* array
# Safe to pass to C functions expecting double*
}
}
