% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dll_loading.R, R/dll_loading_broken.R
\name{dll_compile_and_load}
\alias{dll_compile_and_load}
\title{Compile and load C code dynamically}
\usage{
dll_compile_and_load(
  code,
  name = "temp_dll",
  includes = NULL,
  libs = NULL,
  verbose = FALSE
)

dll_compile_and_load(
  code,
  name = "temp_dll",
  includes = NULL,
  libs = NULL,
  verbose = FALSE
)
}
\arguments{
\item{code}{C code as character string}

\item{name}{Base name for temporary files (default "temp_dll")}

\item{includes}{Additional include directories}

\item{libs}{Additional libraries to link}

\item{verbose}{Print compilation output (default FALSE)}
}
\value{
Library handle that can be used with dll_* functions

Library handle that can be used with dll_* functions
}
\description{
Compile and load C code dynamically

Compile and load C code dynamically
}
