% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_all_offsets}
\alias{ffi_all_offsets}
\title{Get all field offsets for a struct}
\usage{
ffi_all_offsets(struct_type)
}
\arguments{
\item{struct_type}{StructType object}
}
\value{
Named integer vector of offsets
}
\description{
Returns a named integer vector with byte offsets for all fields.
}
\examples{
\dontrun{
Point <- ffi_struct(x = ffi_int(), y = ffi_double())
ffi_all_offsets(Point)
# x y
# 0 8
}

}
