% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-structures.R, R/04-alloc.R
\name{ffi_alloc}
\alias{ffi_alloc}
\title{Allocate memory for FFI structure or array}
\usage{
ffi_alloc(type, ...)

ffi_alloc(type, ...)
}
\arguments{
\item{type}{FFIType object}

\item{n}{Number of elements to allocate (default 1)}
}
\value{
External pointer to buffer
}
\description{
Allocates a buffer for n elements of the given FFIType (e.g., int, double, etc).
Returns an external pointer tagged with the type.
}
