% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_call}
\alias{ffi_call}
\title{Make FFI function call}
\usage{
ffi_call(cif, symbol, ...)
}
\arguments{
\item{cif}{CIF object defining the call interface}

\item{symbol}{NativeSymbol or character name of function}

\item{...}{Arguments to pass to the function (including \code{na_check})}
}
\value{
The return value from the C function, converted to an R type
}
\description{
Call a C function through the FFI interface.
}
\details{
The method implementations accept an additional \code{na_check} argument (logical,
default TRUE). When TRUE, the function checks for NA values in arguments and
errors if found. Set to FALSE to skip NA checking for better performance
(at your own risk).
}
