% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_function}
\alias{ffi_function}
\title{Create a reusable FFI function wrapper}
\usage{
ffi_function(name, return_type, ..., library = NULL)
}
\arguments{
\item{name}{Character name of the function}

\item{return_type}{FFIType for return value}

\item{...}{FFIType objects for arguments}

\item{library}{Character name of library (optional)}
}
\description{
Create a reusable FFI function wrapper
}
