% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R, R/api_mode.R
\name{ffi_get_field}
\alias{ffi_get_field}
\title{Get field value from FFI structure}
\usage{
ffi_get_field(struct_ptr, field_name, struct_type, ...)

ffi_get_field(struct_ptr, field_name, struct_type, ...)
}
\arguments{
\item{struct_ptr}{External pointer to the struct}

\item{field_name}{Name of the field to get (character or integer index)}

\item{struct_type}{Struct type information (list or StructType)}

\item{...}{Additional arguments (not used)}

\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}
}
\value{
Field value

The field value
}
\description{
S7 generic for getting struct field values. Supports both:
\itemize{
\item New API mode: ffi_get_field(ptr, field, struct_type_list)
\item Old StructType: ffi_get_field(ptr, field, StructType)
}
}
\keyword{Types}
