% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_get_field}
\alias{ffi_get_field}
\title{Get field value from FFI structure}
\usage{
ffi_get_field(ptr, field, struct_type, ...)
}
\arguments{
\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}

\item{struct_type}{StructType object}

\item{...}{Additional arguments (not used)}
}
\value{
Field value
}
\description{
Get field value from FFI structure
}
