% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R, R/api_mode.R
\name{ffi_set_field}
\alias{ffi_set_field}
\title{Set field value in FFI structure}
\usage{
ffi_set_field(struct_ptr, field_name, struct_type, value)

ffi_set_field(struct_ptr, field_name, struct_type, value)
}
\arguments{
\item{struct_ptr}{External pointer to the struct}

\item{field_name}{Name of the field to set}

\item{struct_type}{Struct type metadata (contains field offsets and types)}

\item{value}{Value to set}

\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}

\item{...}{Additional arguments (not used)}
}
\value{
Updated pointer

NULL (invisible)
}
\description{
Sets a field value using type-aware writing. Works with bitfields, regular
fields, pointers, nested structs, etc.
}
\keyword{Types}
\keyword{internal}
