% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_set_field}
\alias{ffi_set_field}
\title{Set field value in FFI structure}
\usage{
ffi_set_field(ptr, field, value, struct_type, ...)
}
\arguments{
\item{ptr}{External pointer to structure}

\item{field}{Character field name or integer field index}

\item{value}{Value to set}

\item{struct_type}{StructType object}

\item{...}{Additional arguments (not used)}
}
\value{
Updated pointer
}
\description{
Set field value in FFI structure
}
\section{Types}{
NA
}

