% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions.R
\name{ffi_struct}
\alias{ffi_struct}
\title{Create FFI structure type}
\usage{
ffi_struct(...)
}
\arguments{
\item{...}{Named FFIType objects representing struct fields}
}
\description{
Create FFI structure type
}
