% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{ffi_union}
\alias{ffi_union}
\title{Create FFI union type}
\usage{
ffi_union(...)
}
\arguments{
\item{...}{Named FFIType objects representing union fields}
}
\value{
UnionType object
}
\description{
Create FFI union type
}
\keyword{Types}
