% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{RSimpleFFI-package}
\alias{RSimpleFFI-package}
\title{RSimpleFFI: Simple Foreign Function Interface using 'S7' and 'libffi'}
\description{
Simple Foreign Function Interface for R using libffi and S7 classes. Supports calling C functions with proper type conversion and memory management. Includes standard C types (int8, int16, int32, int64, uint variants), platform types (size_t, bool), and floating point types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/RSimpleFFI}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/RSimpleFFI/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Anthony Green, Red Hat, Inc and others (libffi authors and COPYRIGHT holders) [contributor]
}

}
\keyword{internal}
