% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexp_helpers.R
\name{sexp_ptr}
\alias{sexp_ptr}
\title{Get SEXP pointer from R object with GC protection}
\usage{
sexp_ptr(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
External pointer to the SEXP, with finalizer to release protection
}
\description{
Returns an external pointer to the SEXP (R object header) while ensuring
the object won't be garbage collected as long as the pointer exists.
}
\examples{
\dontrun{
x <- c(1L, 2L, 3L) # Use c() not 1:3 to avoid ALTREP
ptr <- sexp_ptr(x)
# ptr is now safe to pass to C functions expecting SEXP
# When ptr is garbage collected, the protection is released
}
}
