% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{StructType}
\alias{StructType}
\title{FFI Structure Type}
\usage{
StructType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  fields = character(0),
  field_types = list(),
  pack = NULL
)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}

\item{fields}{Character vector of field names}

\item{field_types}{List of FFIType objects for each field}

\item{pack}{Integer packing alignment (NULL for default/natural alignment)}
}
\description{
FFI Structure Type
}
