% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{UnionType}
\alias{UnionType}
\title{FFI Union Type}
\usage{
UnionType(
  name = character(0),
  size = integer(0),
  ref = NULL,
  fields = character(0),
  field_types = list(),
  pack = NULL,
  has_packed_change = logical(0)
)
}
\arguments{
\item{name}{Character name of the type}

\item{size}{Integer size in bytes}

\item{ref}{External pointer to ffi_type}

\item{fields}{Character vector of field names}

\item{field_types}{List of FFIType objects for each field}

\item{pack}{Integer packing alignment (NULL for default/natural alignment)}

\item{has_packed_change}{Logical indicating if packing changes alignment
from natural. When TRUE, the union cannot be passed by value
to C functions (only pointers work) because libffi doesn't support packed unions.}
}
\description{
FFI Union Type
}
